Move conf module into separate standalone package
Making conf module standalone enables using it without previous
initialization (e.g, from main script).
This way, we can run `pytest` directly without conf init via
`runtest.py`, which now can be deprecated.

Resolves #234
