--- conf.py
@@ -1,6 +1,0 @@
-# This module holds all the global.
-
-config = None
-def init(cfg):
-    global config
-    config = cfg
