+++ requirements-dev.txt
@@ -1,2 +1,4 @@
 flake8
 pytest
+coverage
+pytest-cov
