+++ src/modules/hunting/kubelet.py
@@ -6,7 +6,7 @@ import re
 import requests
 import urllib3
 
-from conf import config
+from src.conf import config
 from ...core.events import handler
 from ...core.events.types import Vulnerability, Event, K8sVersionDisclosure
 from ..discovery.kubelet import ReadOnlyKubeletEvent, SecureKubeletEvent
